Remove usage of locale and encoding when gathering libgraphviz info
With LANG=C locale.getdefaultlocale() returns (None, None) which make the code
to retrieve libgraphviz library dirs to fail.

This commit just remove the usage the encoding in a situation when likely is not
needed; this also fails a fail-to-build problem in a minimal chroot environment
(like in Debian packaging system).