Fix segmentation faults in test_clear.py.
On OS X at least, test_clear.py often, but not always segfaulted. Debugging
showed that this was because node_attr and the like were using a
handle value that had already been closed. Now, we update these
attributes to use the current, correct, handle, after it has been
re-opened.