Convert agraph.py warnings from bytes to strings
Warning and error raising behaviour is not correct on Python 3,
where bytes objects cannot be implicitly converted to strings,
and the warnings cause a TypeError crash.

This patch adds default decoding for errors and warnings
that happen when _run_prog is used and fixes errors
on Windows when e.g. no default fonts are available.