Proposal for experimental lib interface.
Add private draw and layout methods that mirror the API of their
public counterparts but are implemented in terms of the Graphviz
library interface.

This solution preserves backwards compatibility while also
incorporating the new interface for preliminary use/testing.

Adds xfailed versions of all tests for draw/layout so that the
experimental methods are tested, though the results will not
cause the suite to fail.