Refactor filehandle type checking.
 * Moves file handle isinstance checks out of swig and into Python
 * Add special case for instance checking of temporary files on
   Windows.

On Windows, tempfile.TemporaryFile objects are not instances of
io.IOBase.
Previously, there was an isinstance check implemented in the graphviz.i
swig wrapper.
This commit moves the isinstance check out of the swig wrapper (where it
is more difficult to grok/modify) and adds the base class of
TemporaryFile on windows to the isinstance check so that it works on all
platforms.