Fix exception on IPython display without layout()
This fixes a regression introduced by the previous commit, where
using IPython to display an AGraph that had no layout would raise an
AttributeError.

With this fix, if layout() has not been called, IPython will just
display a basic text __repr__ of the AGraph; but if layout() has been
called, IPython can display an SVG image.